Adding the ability to filter objects from S3 opendir
This change the listFilter stream context option that can be used when using
opendir and the Amazon S3 stream wrapper. This function can be used to filter
out specific objects from the files yielded from the stream wrapper. This is
utilized in the download sync builder to not attempt to download objects that
have a GLACIER storage class. Closes #322.