Trimming down the number of signed sigv4 headers.
This change updates the SignatureV4 signer to only sign the
"Date", "Host", and "X-Amz-*" headers. This gets around a number
of issues and allows transfer specific things like the Expect
header to be added without causing a signature issue.

I've removed several sigv4 test suite cases that are incompatible
with this change. However, the rest of the test suite and integ
tests are passing.