Update Marshaler.php
Suggestion: Change this method to protected so it can be easily extended. Sometimes in JSON we can have empty strings and an exception will be thrown. In my implementation, it is ok to mark that value as NULL instead. Or, maybe, add a config param to marshalItem/marshalJSON to automatically do this when encountered.