Updating handler list to allow names.
This commit updates the handler list to remove the "sticky" attribute
and instead now has a "name" that can be associated with a middleware by
providing a ":name" after a step (e.g., "init:name"). A step is now
required when calling append and prepend, and you can optionally provide
a name for any middleware. Providing a name allows you to add middleware
before or after another middleware by name or remove a middleware by
name.

In addition to naming middleware, this commit adds more introspection
into the HandlerList: you can now print the list for debugging purposes
and you can provide an interpose function (a function that is triggered
before each middleware is invoked). Interposing is useful for things
like debugging and tracing through a middleware invocation.