Fix listing of root directory of bucket
When I use `scandir('s3://bucket.name/')` the first result is an empty string,
representing the root directory of the bucket. This should rather be a dot.

This is especially required, when using `DirectoryIterator`, as an empty result
will yield this iterator to be empty as well, leading to an empty directory listing on root level.