CloudFront Signer: parse private key once (#1752)
* CloudFront Signer: load private key once

openss_sign() is abysmally slow. On a fast CPU with a small amount of
data it takes 1.5ms of CPU time.

Turns out, half of that was parsing the private key. If we have
openssl parse the key ahead of time, this operation takes half the
time.

This change brought it from 1.5ms to 0.7ms per call, still really
slow but much better.

Note: We started running into performance issues with *lots* of image
urls being rendered on long pages.

* SignerTest: add test for bad private key

* CloudFront Signer Optimization: add changelog entry

* Cloudfront Signer Test: clean up afterward