Fix: warning when passing null as parameter to ucfirst() in Ec2ParamBuilder.php
From src/Api\Serializer/Ec2ParamBuilder.php ucfirst() function is being call with the posibility of passing null as parameter, which could cause a warning if using PHP>=8.1 due that passing null to a function parameter not declared nullable throws a deprecation warning in this php version. To prevent this warning from happening, what I did was to make sure that when calling ucfirst() the variable has value in it.