fix: Warning when symbolic link is pointed to a different path
AssumeRoleWithWebIdentityCredentialProvider try to load the token from a symbolic link, and when this symbolic link is updated and pointed to a path that is different from the previous one, a warning is generated due that is_readable is used to validate if the file exists, and in the first try it has in cache the old token file location, and therefore file_get_contents generate the warning since the token file in cache does not exist. To fix this behavior I removed the is_readable from the logic and I used file_get_contents supressing the warning, which when the file does not exists it will return false, and when the returned value is false the implementation clear the real path cache and another try is made to read the token.