fix: Copy file fails when bucket is an access point arn
S3 client fails to copy when the source bucket is specified through an access point arn, and this is caused because the path of the object specified should contain "/object/" in between the generated bucket url and the object to be copied, and the current implementation does not have this special handling. To fix this behavior I added a condition in the function that generates the source path, where I validate if is an access point then I add the "/object/" part to the path to be returned.