Solve failing unittest CloudFront\SignerTest::testBadPrivateKeyPath()
There were remaining error-messages in the openssl-error-stack, resulting in an
unexpected result in the unittest.

This was solved by outputting all error-messages in Aws\CloudFront\Signer::__construct().

Another option would be to clear all error-messages in
Aws\Test\CloudFront\SignerTest::_tearDown(), so tests would no longer
interfer with eachother.

The first solution was chosen since it would output all
openssl-error-messages to the user, not only the first one,
making it more complete.