fix: standard retry retries on 400 errors
When getting http 40x error and using the standard retry strategy, the 40x error was being retried when this type of error should not be retried, because it could lead to a loop, causing a huge number of retries that will never succeed. This behavior was caused by the way how the conditions were defined in the retry decider logic, that was overridden the decision of not retry this type of error. For reference check and compare the changes.