feat: add straming support on JSON RPC parser
Add event streaming capabilities on the JSON RPC parser, so that operations such as StartLiveTail from CloudwatchLogs are properly handled/parsed. This change also includes an extension of the DecodingEventStreamIterator for handling streamable responses where the stream as response is not seekable, and a handling for when the `event-type` is `initial-response`.