fix: ddb account id metric test (#3051)
Basically, some e2e tests do not use the latest SDK version and therefore latest changes in the endpoint rule set file are not present there, which makes some of tests depending on those latest changes to fail. In this specific case it failed when checking if an account id was resolved or not in dynamodb, because the endpoint ruleset file did not contain the AccountId parameter. For fixing this I use a custom endpoint provider where the ruleset file contains the needed parameter and rule.