fix: closes lazy stream
In a put object request if the body of the request is given with the `SourceFile` parameter, internally there is a middleware that creates a lazy stream based on that file, in order to populate the body of the request, however, this stream is not closed and hence leaving a file open in every request. This change fixes this by making sure the stream is closed once the request is completely sent.