Fix "X-Amz-User-Agent" in blacklist
The header blacklist contains lowercase header names only,
except "X-Amz-User-Agent". This causes this header not to be filtered out,
leading to an error with some proxies that remove the empty header:
> There were headers present in the request which were not signed.

This commit lowercases the header, which correctly removes the header
from the signature.

Related: https://github.com/aws/aws-sdk-php/pull/2259
Resolves: https://github.com/aws/aws-sdk-php/issues/3119